sklearn renamed to scikit-learn
