+++ docs/book/guides/functional-api/functional-api.md
@@ -35,7 +35,7 @@ In order to run the chapters of the guide, you need to install and initialize Ze
 ```bash
 pip install zenml 
 zenml integration install tensorflow
-zenml integration install sklearn
+zenml integration install scikit_learn
 pip install sqlalchemy 
 
 # pull example
