+++ examples/airflow_local/setup.sh
@@ -17,13 +17,13 @@ setup_stack () {
 }
 
 pre_run () {
-  zenml integration install airflow sklearn
+  zenml integration install airflow scikit-learn
 
   setup_stack
 }
 
 pre_run_forced () {
-  zenml integration install airflow sklearn -f
+  zenml integration install airflow scikit-learn -f
 
   setup_stack
 }
