+++ examples/drift_detection/run.py
@@ -17,7 +17,7 @@ import pandas as pd
 from rich import print
 from sklearn import datasets
 
-from zenml.integrations.constants import EVIDENTLY, SKLEARN
+from zenml.integrations.constants import EVIDENTLY, SCIKIT_LEARN
 from zenml.integrations.evidently.steps import (
     EvidentlyProfileConfig,
     EvidentlyProfileStep,
@@ -75,7 +75,7 @@ def analyze_drift(
     return input["data_drift"]["data"]["metrics"]["dataset_drift"]
 
 
-@pipeline(required_integrations=[EVIDENTLY, SKLEARN])
+@pipeline(required_integrations=[EVIDENTLY, SCIKIT_LEARN])
 def drift_detection_pipeline(
     data_loader,
     full_data,
