+++ examples/fetch_historical_runs/README.md
@@ -15,7 +15,7 @@ In order to run this example, you need to install and initialize ZenML:
 pip install zenml
 
 # install ZenML integrations
-zenml integration install sklearn
+zenml integration install scikit_learn
 
 # pull example
 zenml example pull fetch_historical_runs
