+++ examples/fetch_historical_runs/run.py
@@ -15,8 +15,8 @@
 import numpy as np
 from sklearn.base import ClassifierMixin
 
-from zenml.integrations.constants import SKLEARN
-from zenml.integrations.sklearn.helpers.digits import (
+from zenml.integrations.constants import SCIKIT_LEARN
+from zenml.integrations.scikit_learn.helpers.digits import (
     get_digits,
     get_digits_model,
 )
@@ -38,7 +38,7 @@ def trainer(
     X_train: np.ndarray,
     y_train: np.ndarray,
 ) -> ClassifierMixin:
-    """Train a simple sklearn classifier for the digits dataset."""
+    """Train a simple scikit_learn classifier for the digits dataset."""
     model = get_digits_model()
     model.fit(X_train, y_train)
     return model
@@ -70,7 +70,7 @@ def evaluate_and_store_best_model(
     return best_model
 
 
-@pipeline(required_integrations=[SKLEARN])
+@pipeline(required_integrations=[SCIKIT_LEARN])
 def mnist_pipeline(
     importer,
     trainer,
