+++ examples/functional_api/chapter_4.py
@@ -17,7 +17,7 @@ import tensorflow as tf
 from sklearn.base import ClassifierMixin
 from sklearn.linear_model import LogisticRegression
 
-from zenml.integrations.constants import SKLEARN, TENSORFLOW
+from zenml.integrations.constants import SCIKIT_LEARN, TENSORFLOW
 from zenml.pipelines import pipeline
 from zenml.repository import Repository
 from zenml.steps import BaseStepConfig, Output, step
@@ -103,7 +103,7 @@ def sklearn_trainer(
     X_train: np.ndarray,
     y_train: np.ndarray,
 ) -> ClassifierMixin:
-    """Train SVC from sklearn."""
+    """Train SVC from scikit_learn."""
     clf = LogisticRegression(penalty="l1", solver="saga", tol=0.1)
     clf.fit(X_train.reshape((X_train.shape[0], -1)), y_train)
     return clf
@@ -121,7 +121,7 @@ def sklearn_evaluator(
     return test_acc
 
 
-@pipeline(required_integrations=[SKLEARN, TENSORFLOW])
+@pipeline(required_integrations=[SCIKIT_LEARN, TENSORFLOW])
 def mnist_pipeline(
     importer,
     normalizer,
