+++ examples/not_so_quickstart/steps/sklearn_trainer.py
@@ -28,7 +28,7 @@ def sklearn_trainer(
     X_train: np.ndarray,
     y_train: np.ndarray,
 ) -> ClassifierMixin:
-    """Train SVC from sklearn."""
+    """Train SVC from scikit_learn."""
     clf = LogisticRegression(penalty="l1", solver="saga", tol=0.1)
     clf.fit(X_train.reshape((X_train.shape[0], -1)), y_train)
     return clf
