+++ examples/quickstart/setup.sh
@@ -3,9 +3,9 @@
 set -Eeo pipefail
 
 pre_run () {
-  zenml integration install sklearn
+  zenml integration install scikit-learn
 }
 
 pre_run_forced () {
-  zenml integration install sklearn -f
+  zenml integration install scikit-learn -f
 }
