+++ src/zenml/integrations/scikit_learn/helpers/__init__.py
@@ -1,4 +1,4 @@
-#  Copyright (c) ZenML GmbH 2021. All Rights Reserved.
+#  Copyright (c) ZenML GmbH 2022. All Rights Reserved.
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -11,20 +11,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
-from zenml.integrations.constants import SKLEARN
-from zenml.integrations.integration import Integration
-
-
-class SklearnIntegration(Integration):
-    """Definition of sklearn integration for ZenML."""
-
-    NAME = SKLEARN
-    REQUIREMENTS = ["scikit-learn"]
-
-    @classmethod
-    def activate(cls) -> None:
-        """Activates the integration."""
-        from zenml.integrations.sklearn import materializers  # noqa
-
-
-SklearnIntegration.check_installation()
