+++ src/zenml/integrations/scikit_learn/materializers/__init__.py
@@ -11,6 +11,6 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
-from zenml.integrations.sklearn.materializers.sklearn_materializer import (  # noqa
+from zenml.integrations.scikit_learn.materializers.sklearn_materializer import (  # noqa
     SklearnMaterializer,
 )
