+++ src/zenml/integrations/scikit_learn/steps/__init__.py
@@ -12,15 +12,15 @@
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
 
-from zenml.integrations.sklearn.steps.sklearn_evaluator import (
+from zenml.integrations.scikit_learn.steps.sklearn_evaluator import (
     SklearnEvaluator,
     SklearnEvaluatorConfig,
 )
-from zenml.integrations.sklearn.steps.sklearn_splitter import (
+from zenml.integrations.scikit_learn.steps.sklearn_splitter import (
     SklearnSplitter,
     SklearnSplitterConfig,
 )
-from zenml.integrations.sklearn.steps.sklearn_standard_scaler import (
+from zenml.integrations.scikit_learn.steps.sklearn_standard_scaler import (
     SklearnStandardScaler,
     SklearnStandardScalerConfig,
 )
